Fixed bug.

+++ linnea/derivation/graph/base/derivation.py
@@ -88,10 +88,11 @@ class DerivationGraphBase(base.GraphBase):
             else:
                 # Removing existing algorithm files
                 algorithms_dir_name = os.path.join(directory_name, config.language.name, "algorithms")
-                for file in os.listdir(algorithms_dir_name):
-                    path_to_file = os.path.join(algorithms_dir_name, file)
-                    if os.path.isfile(path_to_file):
-                        os.remove(path_to_file)
+                if os.path.exists(algorithms_dir_name):
+                    for file in os.listdir(algorithms_dir_name):
+                        path_to_file = os.path.join(algorithms_dir_name, file)
+                        if os.path.isfile(path_to_file):
+                            os.remove(path_to_file)
 
         if code or pseudocode:
             for n, (path, cost) in enumerate(paths):
